all: cleanup PR

+++ sia/objects/objects.go
@@ -124,7 +124,7 @@ type Snapshot struct {
 	ID          int64
 	CreatedAt   time.Time
 	SiaObjectID types.Hash256
-	ObjectCount int
+	ObjectCount int64
 }
 
 // SnapshotMetadata is attached to a snapshot's Sia object. It lets recovery
